NP-2029 Simplify accordion structure
