+++ src/pages/registration/new_registration/RegistrationAccordion.tsx
@@ -9,6 +9,30 @@ const StyledRegistrationAccordion = styled(Accordion)`
   @media (max-width: ${({ theme }) => `${theme.breakpoints.values.md}px`}) {
     max-width: 90vw;
   }
+
+  .MuiAccordionSummary-content {
+    align-items: center;
+    padding: 1rem 0;
+
+    > :first-child {
+      // Space between icon and title
+      margin-right: 1rem;
+    }
+
+    .MuiSvgIcon-root {
+      font-size: 4rem;
+      @media (max-width: ${({ theme }) => `${theme.breakpoints.values.md}px`}) {
+        display: none;
+      }
+    }
+  }
+
+  .MuiAccordionDetails-root {
+    flex-direction: column;
+    > :not(:first-child) {
+      margin-top: 1rem;
+    }
+  }
 `;
 
 const RegistrationAccordion = ({ children, ...props }: AccordionProps) => (
